Restructuring Express Router
