+++ app.js
@@ -1,14 +1,13 @@
 const express = require("express");
 const app = express();
 const mongoose = require("mongoose");
-const Listing = require("./models/listing.js");
 const path = require("path");
 const methodOverride = require("method-override");
 const ejsMate = require("ejs-mate");
-const wrapAsync = require("./utils/wrapAsync.js");
 const ExpressError = require("./utils/ExpressError.js");
-const { listingSchema, reviewSchema } = require("./schema.js");
-const Review = require("./models/review.js");
+
+const listings = require("./router/listings.js");
+const reviews = require("./router/reviews.js");
 
 app.set("view engine", "ejs");
 app.set("views", path.join(__dirname, "/views"));
@@ -35,132 +34,9 @@ app.get("/", (req, res) => {
   res.send("this is root user");
 });
 
-const validateListing = (req, res, next) => {
-  let { error } = listingSchema.validate(req.body); //Schema Validate for listing
-  if (error) {
-    let errMsg = error.details.map((el) => el.message).join(",");
-    throw new ExpressError(400, errMsg);
-  } else {
-    next();
-  }
-};
+app.use("/listings", listings); //Listing routes
+app.use("/listings/:id/reviews", reviews); //Reviews routes
 
-const validateReview = (req, res, next) => {
-  let { error } = reviewSchema.validate(req.body); //Schema Validate for review
-  if (error) {
-    let errMsg = error.details.map((el) => el.message).join(",");
-    throw new ExpressError(400, errMsg);
-  } else {
-    next();
-  }
-};
-
-//index Route
-app.get(
-  "/listings",
-  wrapAsync(async (req, res) => {
-    let allListings = await Listing.find();
-    res.render("listings/index.ejs", { allListings });
-  })
-);
-
-//New route
-app.get("/listings/new", (req, res) => {
-  res.render("listings/new.ejs");
-});
-
-//Create Route
-app.post(
-  "/listings",
-  validateListing,
-  wrapAsync(async (req, res) => {
-    const newListing = new Listing(req.body.listing);
-    await newListing.save();
-    res.redirect("/listings");
-  })
-);
-
-//Show route
-app.get(
-  "/listings/:id",
-  wrapAsync(async (req, res) => {
-    let { id } = req.params;
-    let listing = await Listing.findById(id).populate("reviews");
-    res.render("listings/show.ejs", { listing });
-  })
-);
-
-//edit route
-app.get(
-  "/listings/:id/edit",
-  wrapAsync(async (req, res) => {
-    let { id } = req.params;
-    const listing = await Listing.findById(id);
-    res.render("listings/edit.ejs", { listing });
-  })
-);
-
-//Update route
-app.put(
-  "/listings/:id",
-  validateListing,
-  wrapAsync(async (req, res) => {
-    let { id } = req.params;
-    await Listing.findByIdAndUpdate(id, { ...req.body.listing });
-    res.redirect(`/listings/${id}`);
-  })
-);
-
-//delete route
-app.delete(
-  "/listings/:id",
-  wrapAsync(async (req, res) => {
-    let { id } = req.params;
-    let DeleteListing = await Listing.findByIdAndDelete(id);
-    console.log(DeleteListing);
-    res.redirect("/listings");
-  })
-);
-
-//Reviews
-//Post Review Route
-app.post(
-  "/listings/:id/reviews",
-  validateReview,
-  wrapAsync(async (req, res) => {
-    let listing = await Listing.findById(req.params.id);
-    let newReview = new Review(req.body.review);
-    listing.reviews.push(newReview);
-    await newReview.save();
-    await listing.save();
-
-    res.redirect(`/listings/${listing._id}`);
-  })
-);
-
-//delete Review route
-app.delete("/listings/:id/reviews/:reviewId", wrapAsync(async (req, res) => {
-  let {id, reviewId} = req.params;
-  await Listing.findByIdAndUpdate(id, {$pull :{reviews: reviewId}});
-  await Review.findByIdAndDelete(reviewId);
-
-  res.redirect(`/listings/${id}`)
-}))
-
-
-// app.get("/testListing", async (req, res) => {
-//   let sampleListing = new Listing({
-//     title: "myHouse",
-//     description: "my Temple like home",
-//     price: 4556,
-//     location: "Dewas Madhaya Pradesh",
-//     country: "India",
-//   });
-
-//   await sampleListing.save();
-//   console.log("sample is saved");
-//   res.send("Okay pass");
-// });
 
 app.listen(8080, () => {
   console.log("Server is listening to port: 8080");
