+++ router/listings.js
@@ -1,0 +1,86 @@
+const express = require("express");
+const router = express.Router();
+const wrapAsync = require("../utils/wrapAsync.js");
+const Listing = require("../models/listing.js");
+const ExpressError = require("../utils/ExpressError.js");
+const { listingSchema } = require("../schema.js");
+
+const validateListing = (req, res, next) => {
+  let { error } = listingSchema.validate(req.body); //Schema Validate for listing
+  if (error) {
+    let errMsg = error.details.map((el) => el.message).join(",");
+    throw new ExpressError(400, errMsg);
+  } else {
+    next();
+  }
+};
+
+//index Route
+router.get(
+  "/",
+  wrapAsync(async (req, res) => {
+    let allListings = await Listing.find();
+    res.render("listings/index.ejs", { allListings });
+  })
+);
+
+//New route
+router.get("/new", (req, res) => {
+  res.render("listings/new.ejs");
+});
+
+//Create Route
+router.post(
+  "/",
+  validateListing,
+  wrapAsync(async (req, res) => {
+    const newListing = new Listing(req.body.listing);
+    await newListing.save();
+    res.redirect("/listings");
+  })
+);
+
+//Show route
+router.get(
+  "/:id",
+  wrapAsync(async (req, res) => {
+    let { id } = req.params;
+    let listing = await Listing.findById(id).populate("reviews");
+    res.render("listings/show.ejs", { listing });
+  })
+);
+
+//edit route
+router.get(
+  "/:id/edit",
+  wrapAsync(async (req, res) => {
+    let { id } = req.params;
+    const listing = await Listing.findById(id);
+    res.render("listings/edit.ejs", { listing });
+  })
+);
+
+//Update route
+router.put(
+  "/:id",
+  validateListing,
+  wrapAsync(async (req, res) => {
+    let { id } = req.params;
+    await Listing.findByIdAndUpdate(id, { ...req.body.listing });
+    res.redirect(`/listings/${id}`);
+  })
+);
+
+//delete route
+router.delete(
+  "/:id",
+  wrapAsync(async (req, res) => {
+    let { id } = req.params;
+    let DeleteListing = await Listing.findByIdAndDelete(id);
+    console.log(DeleteListing);
+    res.redirect("/listings");
+  })
+);
+
+
+module.exports = router;
